Update 02-starting-with-data.Rmd (/issues/576)
The introduction to the data set (line 33) for this lesson did not make much sense to an ecologist. "Species repartition" is not a commonly used phrase when working with animal data, which may have an ecologist confused from the start. I clarified some of the language to include some common phrasing that an ecologist new to R may understand more clearly.

Line 241: c() argument not needed and is confusing when not part of the "Indexing and subsetting data frames" challenge solution with nrow().