add info on submitting issues to contributing.RMD
some contributors add "carpentries checkout contribution" or similar to their issue titles. This is unnecessary clutter in the issues list. I thought it might help to point out in the contribution guidelines that this is not needed.
I also noticed that some issues would be very simple pull requests and would like to encourage contributors to commit their issues as pull requests. I find the Github Desktop app extremely helpful (and fun to use) and think it would be worth mentioning it.